requirements.txt
flask
google-genai
jinja2
fpdf

# fastapi
# uvicorn
pymongo
python-dotenv
